feat: allow rezolus to run on macos (#147)
Allows rezolus to run on macos by scoping all linux-only samplers.
