+++ src/samplers/block_io/linux/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(feature = "bpf")]
+mod bpf;
