+++ src/samplers/block_io/mod.rs
@@ -2,7 +2,7 @@ use crate::*;
 
 sampler!(BlockIO, "block_io", BLOCK_IO_SAMPLERS);
 
-#[cfg(all(feature = "bpf", target_os = "linux"))]
-mod latency;
-
 mod stats;
+
+#[cfg(target_os = "linux")]
+mod linux;
