+++ src/samplers/cpu/linux/mod.rs
@@ -1,0 +1,2 @@
+mod perf;
+mod proc_stat;
