+++ src/samplers/cpu/mod.rs
@@ -5,6 +5,4 @@ sampler!(Cpu, "cpu", CPU_SAMPLERS);
 mod stats;
 
 #[cfg(target_os = "linux")]
-mod perf;
-
-mod proc_stat;
+mod linux;
