+++ src/samplers/memory/linux/mod.rs
@@ -1,0 +1,2 @@
+mod proc_meminfo;
+mod proc_vmstat;
