+++ src/samplers/memory/mod.rs
@@ -5,6 +5,4 @@ sampler!(Memory, "memory", MEMORY_SAMPLERS);
 mod stats;
 
 #[cfg(target_os = "linux")]
-mod proc_meminfo;
-#[cfg(target_os = "linux")]
-mod proc_vmstat;
+mod linux;
