+++ src/samplers/scheduler/mod.rs
@@ -4,5 +4,5 @@ sampler!(Scheduler, "scheduler", SCHEDULER_SAMPLERS);
 
 mod stats;
 
-#[cfg(all(feature = "bpf", target_os = "linux"))]
-mod runqueue;
+#[cfg(target_os = "linux")]
+mod linux;
