+++ src/samplers/tcp/mod.rs
@@ -4,17 +4,5 @@ sampler!(Tcp, "tcp", TCP_SAMPLERS);
 
 mod stats;
 
-mod connection_state;
-mod snmp;
-
-#[cfg(all(feature = "bpf", target_os = "linux"))]
-mod packet_latency;
-
-#[cfg(all(feature = "bpf", target_os = "linux"))]
-mod receive;
-
-#[cfg(all(feature = "bpf", target_os = "linux"))]
-mod retransmit;
-
-#[cfg(all(feature = "bpf", target_os = "linux"))]
-mod traffic;
+#[cfg(target_os = "linux")]
+mod linux;
